requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.63
botocore==1.35.63
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-rest-passwordreset==1.4.2
django-storages==1.14.4
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
packaging==24.1
paypalrestsdk==1.13.3
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.7
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
yarl==1.18.0
